\name{SDPmig-package}
\alias{SDPmig-package}
\alias{SDPmig}
\docType{package}
\title{
\packageTitle{SDPmig}
}
\description{
\packageDescription{SDPmig}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SDPmig}
\packageIndices{SDPmig}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SDPmig}

Maintainer: \packageMaintainer{SDPmig}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
